Created separate files for settings & render, converted class to function component
