+++ src/blocks/social-share/editor.scss
@@ -2,10 +2,7 @@
  * Editor styles for the admin
  */
 
-#wpwrap .edit-post-visual-editor a.uagb-ss__link {
-	color: #3a3a3a;
-}
-
+#wpwrap .edit-post-visual-editor a.uagb-ss__link,
 #wpwrap .edit-post-visual-editor a.uagb-ss__link {
 	color: #3a3a3a;
 }
